# docs/swagger.yaml
definitions:
...
  model.UserJson:
    properties:
      avatar:
        type: string
      captcha:
        type: string
      email:
        type: string
      name:
        type: string
      passWord:
        type: string
      role:
        type: integer
      uid:
        example: "0"
        type: string
      userName:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/seisei/as:
    post:
      parameters:
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取试卷答案
      tags:
      - Teacher
  /api/seisei/mark:
    post:
      parameters:
      - description: 部分试卷ID
        in: formData
        name: partID
        required: true
        type: integer
      - description: 分数
        in: formData
        name: mark
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 为单题打分
      tags:
      - Teacher
  /api/seisei/mission:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      responses: {}
      summary: 获取改卷任务
      tags:
      - Teacher
  /api/seisei/ms/exam:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 查看所在考试
      tags:
      - Teacher
  /api/seisei/ms/fm:
    post:
      parameters:
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 完成阅卷任务
      tags:
      - Teacher
  /api/seisei/ms/part:
    post:
      parameters:
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 按任务阅卷
      tags:
      - Teacher
...
  /api/user/c:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取用户所在的班级
      tags:
      - User
  /api/user/check:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: userName
        required: true
        type: string
      responses: {}
      summary: 检验用户名
      tags:
      - User
  /api/user/info:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 用户信息
      tags:
      - User
  /api/user/list:
    post:
      parameters:
      - description: 用户ID列表
        in: body
        name: IDs
        required: true
        schema:
          items:
            type: string
          type: array
      responses: {}
      summary: 用户信息列表
      tags:
      - User
  /api/user/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: userName
        required: true
        type: string
      - description: 密码
        in: formData
        name: passWord
        required: true
        type: string
      responses: {}
      summary: 登录
      tags:
      - User
  /api/user/new:
    post:
      parameters:
      - description: 用户信息
        in: body
        name: ud
        required: true
        schema:
          $ref: '#/definitions/model.UserJson'
      responses: {}
      summary: 用户注册
      tags:
      - User
  /api/user/rv:
    post:
      parameters:
      - description: 用户信息
        in: body
        name: ud
        required: true
        schema:
          $ref: '#/definitions/model.UserJson'
      responses: {}
      summary: 用户修改信息
      tags:
      - User
  /api/user/verify:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: userName
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      responses: {}
      summary: 用户验证
      tags:
      - User
swagger: "2.0"